+++ src/dirgraphsvg/util/point2d.rs
@@ -1,14 +1,14 @@
 #[derive(PartialEq)]
 pub struct Point2D {
-    pub x: u32,
-    pub y: u32,
+    pub x: i32,
+    pub y: i32,
 }
 
 impl Point2D {
     pub fn move_relative(&self, x: i32, y: i32) -> Self {
         Point2D {
-            x: (self.x as i32 + x) as u32,
-            y: (self.y as i32 + y) as u32,
+            x: self.x + x,
+            y: self.y + y,
         }
     }
 }
